fix: resolve Slack Web API client-side import issues
- Move Slack Web API client to server-side only (/api/slack-notify)
- Update client-side forms to call Slack API via fetch instead of direct import
- Fix Node.js module import errors in Netlify build environment
- Maintain same functionality with proper server/client separation

The @slack/web-api package uses Node.js modules that can't be bundled
for client-side use. This change keeps Slack notifications working
while fixing the build errors.
