+++ app/api/subscribe/route.ts
@@ -1,6 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
 import { z } from "zod";
-import { sendSlackNotification } from "@/lib/slack";
 
 // Request schema
 const subscribeSchema = z.object({
@@ -153,15 +152,25 @@ export async function POST(request: NextRequest) {
       await Promise.allSettled(tagPromises);
     }
 
-    // Send Slack notification
+    // Send Slack notification via internal API
     try {
-      await sendSlackNotification({
-        formType: 'email_subscription',
-        email,
-        name: firstName,
-        source,
-        additionalData: customFields,
+      const slackResponse = await fetch(`${process.env.NEXT_PUBLIC_SITE_URL || 'http://localhost:3000'}/api/slack-notify`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          formType: 'email_subscription',
+          email,
+          name: firstName,
+          source,
+          additionalData: customFields,
+        }),
       });
+      
+      if (!slackResponse.ok) {
+        console.warn('Slack notification failed');
+      }
     } catch (error) {
       console.warn('Failed to send Slack notification:', error);
       // Don't fail the subscription if Slack notification fails
